implemented support for bytecode breakpoints
added notify instruction to asm which is used to tell the sim that a
method has been entered, re-entered (on return from a submethod call)
or exited (when Java execution exits the top level) or that the bc
index has changed. this allows the sim to cache the method and/or bc
idx only when necessary.

modified templateInterpreter to plant notify instructions when running
with product flag -XX:+NotifySimulator (n.b. true by default)

added JNIEXPORT function callback bccheck which is used by the si to
recahce the metod and/or bytecode index.