reinstated fixed that got lost in the merge
n.b. the native signature handler code is installing values relative
to to() == SP not to() == ESP. the SlowSignatureHandler instance does
get passed ESP as the value for pointer _to in its constructor but
only after the generated slow_signature_handler code has set it up to
the location which represents the desired SP value for when we do the
native call. Class SlowSignatureHandler never references to() so there
is not disparity here.