added rest of the int arithmetic ops, optimized L2I conversion, half-added Halt
now have int/long add, sub, multiply, divide, madd, msub working

arithmetic int operations now accept iRegI or iRegL inputs and just
use the bottom 32 bits of the register. IN the latter case this allows
us to avoid generating an explicit l2i (i.e. movw dst src) for the
associated ConvL2I node. This pattern can be used for all int
operations.

unfortunately the matcher will only match the ConvL2I input when it
has 1 use. With multiple uses the ConvL2I is always translated to an
l2i and its destination iReg is supplied as input to the int ops which
consume it. this is still cheap but not as good as it could be.
n.b. Intel always injects the l2i so we are winning anyway (especially
as the single use case will not be uncommon).

implementng idivi/l requires implementation of Halt which should never
get called (it is planted after the bl to the uncommon trap). this
currently plants a brk(999) but needs fixing to generate a proper
trap.

the uncommon trap code for div by 0 still needs testing. it stops with
a haltsim for now.