relocated save and restore of cpu state from doCallout() to caller of run()
n.b. if you pull this change you also need to pull the latest sim tree
(the save and restore in the sim code stop this change fixing the
problem)

the purpose of this change is to ensure the sim print/debug routines
always view the state of the calling JDK code when you are debugging a
call into the JVM.

n.b. the save restore has to be done by the caller of sim.run() after
the caller has retrieved the result in r0/v0. taht is why it cannot be
localized to the sim code itself.